Emit "CATEGORIES" property when building .ics file
The ICS specification includes a "CATEGORIES" property, to specify the
category to which a certain type of event belongs. This can be very
useful in many calendar clients for specifically colouring events of a
certain category, or searching for events by category. Specifically in
the case of working with Durham timetables, it's nice to be able to
highlight tutorials/lectures in different colours, when you're looking
at your week.

This commit adds a field to each activity emitted in the ICS file
generated (when not using the Google integration) which specifies the
category of the event (lecture, tutorial, etc.) based on the
`activity.type` field.